+++ CHANGELOG.md
@@ -4,9 +4,12 @@
 
 ### Breaking
 
+### Changes
+
 ### Feature
 
 - Added basic lazy loading dependencies and boilerplate @sneridagh
+- Added date time widget @nzambello
 
 ### Bugfix